requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4
langchain>=0.2.6
langchain-core>=0.2.10
langchain-openai>=0.1.13
python-dotenv>=1.0.1
